extract the first piece: Pieces into its own file in the folder board_pieces
the __init.py (can be empty) shows that this folder is a module and can be used to import in a python program. Without it, the folder will not be recognized by the python program.

I also refactored in Pieces the methods position and movement so that they do take another argument (board) so that the variable can be used in the functions (since it's not anymore the same file).

IMPROVEMENT: move the variable board (which is now a parameter in position and movement) to the member variables (name, color, col, ...) so that you do not have to use the variable every time you want to call the functions position and movement. Pass the board as an argument when creating a piece (which actually makes sense).